Added some optimizations to ForEachExpression.
An enumerator is no longer used for array sequences.

If a public instance method "GetEnumerator()" returning a compatible
enumerator is found, that method will be called regardless of whether it
overrides IEnumerable.GetEnumerator().  These are the same semantics as
the C# 'foreach' loop and enables boxing avoidance for value type
enumerators.

If an enumerator is known to implement IDisposable, or known *not* to
implement it, then the type check or Dispose() call is omitted,
respectively.